Fixed runConsensusRegions returning NULL for a CompressedGRangesList
The GenomicRanges Package has been updated since this package was last compiled.  The GRangesList() function now returns a CompressedGRangesList object instead of a GRangesList object in many cases.  This commit adds CompressedGRangesList as an acceptable object for the runConsensusRegions function.

Also updated function to produce an appropriate error message if the input is incorrect instead of silently retruning NULL